supplicant:  Allow configuring percentage for ignore_auth_resp
This lets one configure a station to drop a percentage
(including 0 and 100%) of the RX management frames to
aid testing applications.

The value is moved to the config section so that it can be
defined in a supplicant config file.  There is a functional
change in that before you could reset a station through
wpa_cli and it would clear the ignore_auth_resp value,
and now you would have to explicitly clear it since it is
a configuration.

To enable the feature, you must compile with CONFIG_TESTING_OPTIONS
enabled.

Signed-off-by: Ben Greear <greearb@candelatech.com>