supplicant: Support disabling the supported-oper-class IE.
This provides a work-around for cases where the AP (most likely)
has bugs when this element exists and/or has 5Ghz elements when
the AP is evidently unexpecting that.

Test case that required this fix is a DPC3941 AP, configured with
2.4G radio to have WPA2 PSK.  Client station is ath10k 9880.
The symptom is that the client radio does not ack the 1/4 eapol.
I do not know exactly why, but maybe it is corrupted in some suble
way that I did not notice.

Signed-off-by: Ben Greear <greearb@candelatech.com>