supplicant: Tune auto-roaming based on scan results, allow configuring thresholds.
While trying to force an .11r roam, I noticed that when I scanned, the station
went back to the original AP.  The RSSI was around -25 for one and -35 for
the other, but the RSSI threshold was 5, so it flipped.

I believe that the performance difference between RSSI 25 and 35 is negligible,
so I modified the algorithm to basically not auto-roam if both signals are
good.  I'm not certain my new automatic thresholds are perfect, so maybe
they need more tuning.  At the least, we should not auto roam between an RSSI 25
and RSSI 35 AP based on RSSI.

And, allow user overrides in case we want the thresholds even higher.

In addition, add a note regarding the estimated bandwidth calculation,
as it does not appear to take NSS into account.

Signed-off-by: Ben Greear <greearb@candelatech.com>