osu:  Add timeout before we check for IP, slurp event buffer.
In my setup, it was taking a small bit of time to remove the IP
address due to calling code to do a dhcp release instead of just
removing the IP.  So, add a 5 second timeout before we start looking
for the new IP.

Also, read from the event socket before doing an action and checking
the real event.  I was suspicios that maybe an old event text could
be in there to give a false positive, but no idea if that is really happening
or not.

Signed-off-by: Ben Greear <greearb@candelatech.com>