hostapd:  Allow configuring local-bss for RNR in config file.
For cases where one does not want to use multiple bss in one hostapd
instance, allow configuring a hostapd instance to advertise RNR
for a separate instance.

Code complexity makes it hard to support the old and the new RNR
bss logic when building IEs, so for now, if local_bss is configured
in the config file, then multiple bss will cause RNR to be enabled.

Signed-off-by: Ben Greear <greearb@candelatech.com>